+++ engine/src/org/pentaho/di/trans/steps/checksum/CheckSum.java
@@ -187,7 +187,7 @@ public class CheckSum extends BaseStep implements StepInterface {
 		StringBuffer sb = new StringBuffer();
 		for (int i = 0; i < bytes.length; i++) {
 			byte b = bytes[i];
-			sb.append((int) (0x00FF & b));
+			sb.append(0x00FF & b);
 			if (i + 1 < bytes.length) {
 				sb.append("-");
 			}
